refactor: extract search partition planning
