+++ src/search/src/lib.rs
@@ -18,6 +18,7 @@ pub mod datafusion;
 pub mod enrichment;
 pub mod index;
 pub mod inspector;
+pub mod partition;
 pub mod sql;
 pub mod tantivy;
 pub mod udf;
