+++ src/search/src/partition/regular_partition.rs
